+++ docs/changelog.rst
@@ -1,9 +1,14 @@
 Changelog
 ~~~~~~~~~
 
+0.2.0rc15 (July 2023)
+---------------------
+* additional optimisation for cleaning up empty blocks
+
 0.2.0rc14 (July 2023)
 ---------------------
 * add simplification for RangePredicate in case of equal bounds
+* allow conditional circboxes in the circuit
 * update pytket requirement to 1.17
 
 0.2.0rc13 (June 2023)